~:projects:air-quality:/air-quality-raiym/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
paho-mqtt==1.6.1
requests==2.31.0
pytest==7.4.0
